+++ civic_digital_twins/__init__.py
@@ -1,0 +1,1 @@
+"""Civic Digital Twins Modelling Framework."""
