+++ civic_digital_twins/dt_model/__init__.py
@@ -1,16 +1,16 @@
 """The dt_model package implements the tool for digital twins modeling and simulation."""
 
-from dt_model.model.model import Model
-from dt_model.simulation.ensemble import Ensemble
-from dt_model.symbols.constraint import Constraint
-from dt_model.symbols.context_variable import (
+from .model.model import Model
+from .simulation.ensemble import Ensemble
+from .symbols.constraint import Constraint
+from .symbols.context_variable import (
     CategoricalContextVariable,
     ContextVariable,
     ContinuousContextVariable,
     UniformCategoricalContextVariable,
 )
-from dt_model.symbols.index import ConstIndex, Index, LognormDistIndex, SymIndex, TriangDistIndex, UniformDistIndex
-from dt_model.symbols.presence_variable import PresenceVariable
+from .symbols.index import ConstIndex, Index, LognormDistIndex, SymIndex, TriangDistIndex, UniformDistIndex
+from .symbols.presence_variable import PresenceVariable
 
 __all__ = [
     "CategoricalContextVariable",
