+++ civic_digital_twins/dt_model/internal/sympyke/piecewise.py
