+++ civic_digital_twins/dt_model/model/abstract_model.py
@@ -2,10 +2,10 @@
 
 from __future__ import annotations
 
-from dt_model.symbols.constraint import Constraint
-from dt_model.symbols.context_variable import ContextVariable
-from dt_model.symbols.index import Index
-from dt_model.symbols.presence_variable import PresenceVariable
+from ..symbols.constraint import Constraint
+from ..symbols.context_variable import ContextVariable
+from ..symbols.index import Index
+from ..symbols.presence_variable import PresenceVariable
 
 
 class AbstractModel:
