+++ civic_digital_twins/dt_model/simulation/ensemble.py
@@ -4,9 +4,9 @@ from __future__ import annotations
 
 from functools import reduce
 
-from dt_model.internal.sympyke.symbol import SymbolValue
-from dt_model.model.instantiated_model import InstantiatedModel
-from dt_model.symbols.context_variable import ContextVariable
+from ..internal.sympyke.symbol import SymbolValue
+from ..model.instantiated_model import InstantiatedModel
+from ..symbols.context_variable import ContextVariable
 
 
 class Ensemble:
