+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "civic-digital-twins"
-version = "0.3.0"
+version = "0.4.0"
 description = "Civic-Digital-Twins Modeling Framework"
 readme = "README.md"
 authors = [
@@ -36,12 +36,12 @@ dev = [
 Homepage = "https://github.com/fbk-most/dt-model"
 
 [tool.hatch.build.targets.wheel]
-packages = ["dt_model"]
+packages = ["civic_digital_twins"]
 
 [tool.pytest.ini_options]
 testpaths = ["tests"]
 python_files = ["test_*.py"]
-addopts = "--cov=dt_model --cov-report=xml --cov-report=term-missing"
+addopts = "--cov=civic_digital_twins --cov-report=xml --cov-report=term-missing"
 
 [tool.ruff]
 target-version = "py311"
