+++ tests/engine/numpybackend/test_executor.py
@@ -5,8 +5,8 @@
 import numpy as np
 import pytest
 
-from dt_model.engine.frontend import graph, linearize
-from dt_model.engine.numpybackend import executor
+from civic_digital_twins.dt_model.engine.frontend import graph, linearize
+from civic_digital_twins.dt_model.engine.numpybackend import executor
 
 
 def test_constant_evaluation():
