+++ tests/examples/molveno/test_overtourism.py
@@ -4,8 +4,8 @@ import random
 
 import numpy as np
 
-from dt_model import Constraint, ContextVariable, Ensemble
-from dt_model.examples.molveno.overtourism import (
+from civic_digital_twins.dt_model import Constraint, ContextVariable, Ensemble
+from civic_digital_twins.dt_model.examples.molveno.overtourism import (
     CV_season,
     CV_weather,
     CV_weekday,
@@ -13,8 +13,8 @@ from dt_model.examples.molveno.overtourism import (
     PV_excursionists,
     PV_tourists,
 )
-from dt_model.internal.sympyke.symbol import Symbol, SymbolValue
-from dt_model.model.instantiated_model import InstantiatedModel
+from civic_digital_twins.dt_model.internal.sympyke.symbol import Symbol, SymbolValue
+from civic_digital_twins.dt_model.model.instantiated_model import InstantiatedModel
 
 
 def compare_constraint_results(got: dict[Constraint, np.ndarray], expect: dict[str, np.ndarray]) -> list[str]:
