+++ tests/internal/sympyke/test_operators.py
@@ -4,9 +4,9 @@
 
 import numpy as np
 
-from dt_model.engine.frontend import graph, linearize
-from dt_model.engine.numpybackend import executor
-from dt_model.internal.sympyke import Eq, Symbol
+from civic_digital_twins.dt_model.engine.frontend import graph, linearize
+from civic_digital_twins.dt_model.engine.numpybackend import executor
+from civic_digital_twins.dt_model.internal.sympyke import Eq, Symbol
 
 
 def test_eq_basics():
